email_agent/requirements.txt
notion-client>=2.2.1
openai>=1.93.0
python-dotenv>=1.0.0
pytz>=2024.1